refactor httpx instrumentation
httpx 0.21 and httpcore 0.14 once again changed how it works,
necessitating more special handling. To make this a tad less
cumbersome, most of this code has moved into a utils file.
To make the structure somewhat sane, httpx got its own
package in instrumentation
