+++ .ci/.jenkins_framework_full.yml
@@ -77,6 +77,7 @@ FRAMEWORK:
   - graphene-2
   - httpx-0.13
   - httpx-0.14
+  - httpx-0.21
   - httpx-newest
   - httplib2-newest
   - sanic-20.12
